backend/pyproject.toml
[project]
name = "estatefox-api"
version = "0.1.0"
description = "A full-stack real estate application for South Florida property listings, search, and management."
authors = [{ name = "Developer", email = "dev@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "mangum>=0.17.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.34.0",
    "pyhumps>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "moto>=5.0.0",
    "uvicorn>=0.27.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "boto3-stubs[bedrock-runtime,ses,s3]>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/estatefox_api"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
